# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Shop API",
        "description": "Shop API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Barcha mahsulotlar ro'yxati",
                "operationId": "7cd9d394f726672223f216b77acab086",
                "responses": {
                    "200": {
                        "description": "Mahsulotlar muvaffaqiyatli qaytarildi"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Yangi mahsulot yaratish",
                "operationId": "e83af3dd12f9ebc4ad755304893c016f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "MacBook M4"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Yangi M4 chipli MacBook"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 3000
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "products/macbook.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mahsulot yaratildi"
                    },
                    "422": {
                        "description": "Validatsiya xatoliklari"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Bitta mahsulotni ko‘rish",
                "operationId": "fa07b5c9bbd4d51f1f3763ca56e0908b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mahsulot topildi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Mahsulotni yangilash",
                "operationId": "e0e4bcc68686fa05279e23580e41bea1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "MacBook Pro"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Yangilangan model"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 3500
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "products/macbook-pro.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mahsulot yangilandi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Mahsulotni o‘chirish",
                "operationId": "bae80a1b7d88ec2b8f025a048aa00f86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "O‘chirildi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/search": {
            "get": {
                "tags": [
                    "Mahsulotlar"
                ],
                "summary": "Mahsulotlar ro'yxatini olish (qidirish bilan)",
                "operationId": "760b379b48484c90d73af2ccc9f8ba9d",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Mahsulot nomi yoki tavsifi bo‘yicha qidirish",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mahsulotlar ro'yxati",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "iPhone 13"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Eng yangi iPhone modeli"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 999.99
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "products/iphone13.jpg"
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-06-15T12:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-06-15T12:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}